Add files via upload
This is the main program set. Here, the user inputs the movie's script and cast list (via IMDB.com) into two directories. The script must be named after the movie, and the cast list is named after the movie plus "Cast." The program prompts the user for the two directories, and then locates the two files by their matched name. The program reads the script line by line, and determines who the characters might be according to the lines spoken. Using the cast list, the program then matches the actors to the potential characters found in the script. At that point the program has linked the actor to the character, and the character to their lines. The data is then exported to a joint excel file for further processing.